Conflicts:
	crawler/profiles/integracion/crawler.core.properties
	crawler/src/main/java/es/inteco/crawler/job/CrawlerJob.java
	crawler/src/main/java/es/inteco/utils/MailUtils.java
	intavcore/profiles/desarrollo/intav.properties
	intavcore/profiles/integracion/intav.properties
	portal/profiles/desarrollo/context.xml
	portal/profiles/desarrollo/crawler.properties
	portal/profiles/desarrollo/pdf.properties
	portal/profiles/integracion/context.xml
	portal/profiles/integracion/crawler.properties
	portal/profiles/integracion/log4j.properties
	portal/profiles/integracion/pdf.properties
	portal/src/main/java/es/inteco/rastreador2/pdf/utils/PDFUtils.java
	portal/src/main/java/es/inteco/rastreador2/utils/basic/service/BasicServiceUtils.java
	portal/src/main/resources/ApplicationResources.properties
	portal/src/main/resources/ApplicationResources_es.properties
	portal/src/main/webapp/common/crawler_header.jsp
	portal/src/main/webapp/common/crawler_login_header.jsp
	scripts/OAW.sql

+++ portal/src/main/java/es/inteco/rastreador2/action/basic/service/BasicServiceAction.java
@@ -3,14 +3,16 @@ package es.inteco.rastreador2.action.basic.service;
 import es.inteco.common.Constants;
 import es.inteco.common.logging.Logger;
 import es.inteco.common.properties.PropertiesManager;
+import es.inteco.common.utils.StringUtils;
 import es.inteco.crawler.job.CrawledLink;
 import es.inteco.crawler.job.CrawlerData;
 import es.inteco.crawler.job.CrawlerJob;
-import es.inteco.intav.utils.StringUtils;
 import es.inteco.plugin.dao.DataBaseManager;
 import es.inteco.rastreador2.actionform.basic.service.BasicServiceForm;
 import es.inteco.rastreador2.dao.basic.service.DiagnosisDAO;
 import es.inteco.rastreador2.pdf.BasicServiceExport;
+import es.inteco.rastreador2.pdf.builder.AnonymousResultExportPdfUNE2004;
+import es.inteco.rastreador2.pdf.builder.AnonymousResultExportPdfUNE2012;
 import es.inteco.rastreador2.pdf.utils.PrimaryExportPdfUtils;
 import es.inteco.rastreador2.utils.CrawlerUtils;
 import es.inteco.rastreador2.utils.basic.service.BasicServiceConcurrenceSystem;
@@ -34,15 +36,16 @@ import java.util.Date;
 import java.util.Iterator;
 import java.util.List;
 
+import static es.inteco.common.Constants.CRAWLER_CORE_PROPERTIES;
 import static es.inteco.common.Constants.CRAWLER_PROPERTIES;
 
 public class BasicServiceAction extends Action {
 
     @Override
-    public ActionForward execute(ActionMapping mapping, ActionForm form, HttpServletRequest request, HttpServletResponse response)
+    public ActionForward execute(final ActionMapping mapping, final ActionForm form, final HttpServletRequest request, final HttpServletResponse response)
             throws Exception {
 
-        String action = request.getParameter(Constants.ACTION);
+        final String action = request.getParameter(Constants.ACTION);
 
         if (!StringUtils.isEmpty(action) && action.equalsIgnoreCase(Constants.EXECUTE)) {
             executeCrawling(form, request);
@@ -54,14 +57,14 @@ public class BasicServiceAction extends Action {
     }
 
 
-    private ActionForward launchCrawling(ActionMapping mapping, ActionForm form, HttpServletRequest request, HttpServletResponse response) throws Exception {
-        BasicServiceForm basicServiceForm = BasicServiceUtils.getBasicServiceForm((BasicServiceForm) form, request);
+    private ActionForward launchCrawling(final ActionMapping mapping, final ActionForm form, final HttpServletRequest request, final HttpServletResponse response) throws Exception {
+        final BasicServiceForm basicServiceForm = BasicServiceUtils.getBasicServiceForm((BasicServiceForm) form, request);
 
         // Guardamos en base de datos y en los logs las peticiones de análisis
         Logger.putLog("El usuario " + basicServiceForm.getUser() + " ha lanzado una petición de análisis en el servicio básico", BasicServiceAction.class, Logger.LOG_LEVEL_INFO);
 
         basicServiceForm.setDomain(BasicServiceUtils.checkIDN(basicServiceForm.getDomain()));
-        ActionErrors errors = basicServiceForm.validate(mapping, request);
+        final ActionErrors errors = basicServiceForm.validate(mapping, request);
 
         if (errors.isEmpty()) {
             errors.add(BasicServiceUtils.validateReport(basicServiceForm));
@@ -77,7 +80,7 @@ public class BasicServiceAction extends Action {
             basicServiceForm.setId(BasicServiceUtils.saveRequestData(basicServiceForm, Constants.BASIC_SERVICE_STATUS_QUEUED));
 
             // Ponemos el análisis en cola
-            BasicServiceQueingThread basicServiceQueingThread = new BasicServiceQueingThread(basicServiceForm);
+            final BasicServiceQueingThread basicServiceQueingThread = new BasicServiceQueingThread(basicServiceForm);
             basicServiceQueingThread.start();
 
             CrawlerUtils.returnText(getResources(request).getMessage("basic.service.queued"), response, false);
@@ -85,15 +88,15 @@ public class BasicServiceAction extends Action {
             basicServiceForm.setId(BasicServiceUtils.saveRequestData(basicServiceForm, Constants.BASIC_SERVICE_STATUS_LAUNCHED));
 
             // Lanzamos el análisis
-            BasicServiceThread basicServiceThread = new BasicServiceThread(basicServiceForm);
+            final BasicServiceThread basicServiceThread = new BasicServiceThread(basicServiceForm);
             basicServiceThread.start();
 
             CrawlerUtils.returnText(getResources(request).getMessage("basic.service.launched"), response, false);
         } else {
-            StringBuilder text = new StringBuilder(getResources(request).getMessage("basic.service.validation.errors"));
+            final StringBuilder text = new StringBuilder(getResources(request).getMessage("basic.service.validation.errors"));
 
             for (Iterator<ActionMessage> iterator = errors.get(); iterator.hasNext(); ) {
-                ActionMessage message = iterator.next();
+                final ActionMessage message = iterator.next();
                 text.append("\n - ").append(getResources(request).getMessage(message.getKey(), message.getValues()));
             }
 
@@ -105,11 +108,10 @@ public class BasicServiceAction extends Action {
         return null;
     }
 
-    private ActionForward executeCrawling(ActionForm form, HttpServletRequest request) throws Exception {
-        PropertiesManager pmgr = new PropertiesManager();
+    private ActionForward executeCrawling(final ActionForm form, final HttpServletRequest request) throws Exception {
+        Logger.putLog("executeCrawling", BasicServiceAction.class, Logger.LOG_LEVEL_DEBUG);
 
         Connection conn = null;
-        Connection c = null;
 
         String pdfPath = null;
 
@@ -122,15 +124,12 @@ public class BasicServiceAction extends Action {
 
         try {
             conn = DataBaseManager.getConnection();
-            c = DataBaseManager.getConnection(pmgr.getValue(CRAWLER_PROPERTIES, "datasource.name.intav"));
             //Lanzamos el rastreo de INTAV
-            CrawlerJob crawlerJob = new CrawlerJob();
+            final CrawlerJob crawlerJob = new CrawlerJob();
+            final Long idCrawling = System.nanoTime() * (-1);
+            final CrawlerData crawlerData = createCrawlerData(basicServiceForm, BasicServiceUtils.getGuideline(basicServiceForm.getReport()), idCrawling);
 
-            Long idCrawling = System.nanoTime() * (-1);
-
-            CrawlerData crawlerData = createCrawlerData(basicServiceForm, BasicServiceUtils.getGuideline(basicServiceForm.getReport()), idCrawling);
-
-            List<CrawledLink> crawledLinks;
+            final List<CrawledLink> crawledLinks;
             if (crawlerData.getUrls() != null && !crawlerData.getUrls().isEmpty()) {
                 crawledLinks = crawlerJob.testCrawler(crawlerData);
             } else {
@@ -138,39 +137,43 @@ public class BasicServiceAction extends Action {
             }
 
             if (!crawledLinks.isEmpty()) {
-                DateFormat df = new SimpleDateFormat(pmgr.getValue(CRAWLER_PROPERTIES, "file.date.format"));
+                final PropertiesManager pmgr = new PropertiesManager();
+                final DateFormat df = new SimpleDateFormat(pmgr.getValue(CRAWLER_PROPERTIES, "file.date.format"));
 
                 pdfPath = pmgr.getValue(CRAWLER_PROPERTIES, "pdf.basic.service.path") + idCrawling + File.separator + basicServiceForm.getName() + "_" + df.format(new Date()) + ".pdf";
 
                 //Creamos el PDF con los resultados del rastreo
-                List<Long> evaluationIds = DiagnosisDAO.getEvaluationIds(idCrawling);
+                final List<Long> evaluationIds = DiagnosisDAO.getEvaluationIds(idCrawling);
                 if (basicServiceForm.getReport().equalsIgnoreCase(Constants.REPORT_UNE) ||
                         basicServiceForm.getReport().equalsIgnoreCase(Constants.REPORT_UNE_FILE) ||
                         basicServiceForm.getReport().equalsIgnoreCase(Constants.REPORT_WCAG_1_FILE) ||
                         basicServiceForm.getReport().equalsIgnoreCase(Constants.REPORT_WCAG_2_FILE)) {
                     BasicServiceExport.generatePDF(request, basicServiceForm, evaluationIds, idCrawling, pdfPath);
                 } else if (basicServiceForm.getReport().equals(Constants.REPORT_OBSERVATORY) ||
-                        basicServiceForm.getReport().equals(Constants.REPORT_OBSERVATORY_FILE)) {
-                    String content = basicServiceForm.isContentAnalysis() ? basicServiceForm.getContent() : null;
-                    PrimaryExportPdfUtils.exportToPdf(idCrawling, evaluationIds, request, pdfPath, basicServiceForm.getName(), content, -System.currentTimeMillis(), 1);
+                        basicServiceForm.getReport().equals(Constants.REPORT_OBSERVATORY_FILE) || basicServiceForm.getReport().equals(Constants.REPORT_OBSERVATORY_1_NOBROKEN)) {
+                    Logger.putLog("Exportando desde BasicService a PrimaryExportPdfUtils.exportToPdf(new AnonymousResultExportPdfUNE2004() ...", BasicServiceAction.class, Logger.LOG_LEVEL_DEBUG);
+                    final String content = basicServiceForm.isContentAnalysis() ? basicServiceForm.getContent() : null;
+                    PrimaryExportPdfUtils.exportToPdf(new AnonymousResultExportPdfUNE2004(), idCrawling, evaluationIds, request, pdfPath, basicServiceForm.getName(), content, -System.currentTimeMillis(), 1);
+                } else if (Constants.REPORT_OBSERVATORY_2.equals(basicServiceForm.getReport()) || Constants.REPORT_OBSERVATORY_2_NOBROKEN.equals(basicServiceForm.getReport())) {
+                    Logger.putLog("Exportando desde BasicService a PrimaryExportPdfUtils.exportToPdf(new AnonymousResultExportPdfUNE2012() ...", BasicServiceAction.class, Logger.LOG_LEVEL_DEBUG);
+                    final String content = basicServiceForm.isContentAnalysis() ? basicServiceForm.getContent() : null;
+                    PrimaryExportPdfUtils.exportToPdf(new AnonymousResultExportPdfUNE2012(), idCrawling, evaluationIds, request, pdfPath, basicServiceForm.getName(), content, -System.currentTimeMillis(), 1);
                 }
 
                 // Comprimimos el fichero
                 pdfPath = BasicServiceExport.compressReport(pdfPath);
-
                 //Borramos el análisis
-                DiagnosisDAO.deleteAnalysis(c, basicServiceForm.getName(), idCrawling);
+                DiagnosisDAO.deleteAnalysis(conn, basicServiceForm.getName(), idCrawling);
 
                 //Lo enviamos por correo electrónico
-                String subject = basicServiceForm.getReport().equals(Constants.REPORT_OBSERVATORY) ? getResources(request).getMessage("basic.service.mail.subject.observatory") : getResources(request).getMessage("basic.service.mail.subject.une");
-                String text = getMailText(request, basicServiceForm);
-                ArrayList<String> mailTo = new ArrayList<String>();
+                final String subject = getMailSubject(basicServiceForm.getReport());
+                final String text = getMailText(request, basicServiceForm);
+                final ArrayList<String> mailTo = new ArrayList<String>();
                 mailTo.add(basicServiceForm.getEmail());
-                String replyTo = pmgr.getValue(CRAWLER_PROPERTIES, "basic.service.reply.to.mail");
-                String replyToName = pmgr.getValue(CRAWLER_PROPERTIES, "basic.service.reply.to.name");
-
-                MailUtils.sendMail(pmgr.getValue(CRAWLER_PROPERTIES, "basic.service.address"), pmgr.getValue(CRAWLER_PROPERTIES, "basic.service.name"),
-                        mailTo, subject, text, pdfPath, new File(pdfPath).getName(), replyTo, replyToName, false);
+                final String mailFrom = pmgr.getValue(CRAWLER_CORE_PROPERTIES, "mail.address.from");
+                Logger.putLog("Enviando correo del servicio de diagnóstico", BasicServiceAction.class, Logger.LOG_LEVEL_INFO);
+                MailUtils.sendMail(mailFrom, "Servicio on-line de diagnóstico de Accesibilidad",
+                        mailTo, subject, text, pdfPath, new File(pdfPath).getName(), mailFrom, "Servicio on-line de diagnóstico de Accesibilidad", false);
                 BasicServiceUtils.updateRequestStatus(basicServiceForm, Constants.BASIC_SERVICE_STATUS_FINISHED);
             } else {
                 // Avisamos de que ha sido imposible acceder a la página a rastrear
@@ -193,9 +196,8 @@ public class BasicServiceAction extends Action {
             BasicServiceUtils.updateRequestStatus(basicServiceForm, Constants.BASIC_SERVICE_STATUS_ERROR);
         } finally {
             DataBaseManager.closeConnection(conn);
-            DataBaseManager.closeConnection(c);
 
-            if (StringUtils.isNotEmpty(pdfPath)) {
+            if (pdfPath != null && StringUtils.isNotEmpty(pdfPath)) {
                 // Borramos la carpeta con el PDF generado
                 File pdfFile = new File(pdfPath);
                 FileUtils.deleteDir(pdfFile.getParentFile());
@@ -205,10 +207,9 @@ public class BasicServiceAction extends Action {
         return null;
     }
 
-    private CrawlerData createCrawlerData(BasicServiceForm basicServiceForm, Long idGuideline, Long idCrawling) {
-        CrawlerData crawlerData = new CrawlerData();
-        String[] cartuchos = new String[1];
-        cartuchos[0] = "es.inteco.accesibilidad.CartuchoAccesibilidad";
+    private CrawlerData createCrawlerData(final BasicServiceForm basicServiceForm, final Long idGuideline, final Long idCrawling) {
+        final CrawlerData crawlerData = new CrawlerData();
+        final String[] cartuchos = new String[]{"es.inteco.accesibilidad.CartuchoAccesibilidad"};
         crawlerData.setCartuchos(cartuchos);
         crawlerData.setIdCrawling(idCrawling);
         crawlerData.setIdFulfilledCrawling(idCrawling);
@@ -218,36 +219,70 @@ public class BasicServiceAction extends Action {
         crawlerData.setPseudoaleatorio(true);
         crawlerData.setTopN(Integer.parseInt(basicServiceForm.getAmplitud()));
         crawlerData.setUser(basicServiceForm.getUser());
-        ArrayList<String> mails = new ArrayList<String>();
+        final ArrayList<String> mails = new ArrayList<String>();
         mails.add(basicServiceForm.getEmail());
         crawlerData.setUsersMail(mails);
         crawlerData.setTest(true);
         if (StringUtils.isNotEmpty(basicServiceForm.getDomain())) {
-            List<String> urls = new ArrayList<String>();
+            final List<String> urls = new ArrayList<String>();
             urls.add(basicServiceForm.getDomain());
             crawlerData.setUrls(urls);
         } else {
             crawlerData.setContent(basicServiceForm.getContent());
         }
-        crawlerData.setFicheroNorma(CrawlerUtils.getFicheroNorma(idGuideline));
+
+        crawlerData.setFicheroNorma(includeBrokenLinksCheck(CrawlerUtils.getFicheroNorma(idGuideline), basicServiceForm.getReport()));
         crawlerData.setDomains(es.inteco.utils.CrawlerUtils.addDomainsToList(basicServiceForm.getDomain(), true, Constants.ID_LISTA_SEMILLA));
         crawlerData.setInDirectory(basicServiceForm.isInDirectory());
-
+        Logger.putLog(crawlerData.getFicheroNorma(), BasicServiceAction.class, Logger.LOG_LEVEL_DEBUG);
         return crawlerData;
     }
 
-    private String getMailText(HttpServletRequest request, BasicServiceForm basicServiceForm) {
-        String text;
+    private String includeBrokenLinksCheck(final String ficheroNorma, final String report) {
+        if (report.endsWith("-nobroken")) {
+            return ficheroNorma.substring(0, ficheroNorma.length() - 4) + "-nobroken.xml";
+        } else {
+            return ficheroNorma;
+        }
+    }
+
+    private String getMailSubject(final String reportType) {
+        if (Constants.REPORT_OBSERVATORY.equals(reportType) || Constants.REPORT_OBSERVATORY_FILE.equals(reportType) || Constants.REPORT_OBSERVATORY_1_NOBROKEN.equals(reportType)) {
+            return "Informe de Accesibilidad Web: Observatorio UNE 2004";
+        } else if (Constants.REPORT_OBSERVATORY_2.equals(reportType) || Constants.REPORT_OBSERVATORY_2_NOBROKEN.equals(reportType)) {
+            return "Informe de Accesibilidad Web: Observatorio UNE 2012";
+        } else if ("une".equals(reportType)) {
+            return "Informe de Accesibilidad Web: UNE 139803";
+        }
+        return "Informe de Accesibilidad Web";
+    }
+
+    private String getMailText(final HttpServletRequest request, final BasicServiceForm basicServiceForm) {
+        final String text;
         if (basicServiceForm.isContentAnalysis()) {
-            String[] args = {basicServiceForm.getUser(), BasicServiceUtils.getTitleFromContent(basicServiceForm.getContent()), basicServiceForm.getProfundidad(), basicServiceForm.getAmplitud(), basicServiceForm.getReport()};
+            final String[] args = {basicServiceForm.getUser(), BasicServiceUtils.getTitleFromContent(basicServiceForm.getContent()), basicServiceForm.getProfundidad(), basicServiceForm.getAmplitud(), reportToString(basicServiceForm.getReport())};
             text = basicServiceForm.getReport().equals(Constants.REPORT_OBSERVATORY) ? getResources(request).getMessage("basic.service.mail.text.observatory.content", args) : getResources(request).getMessage("basic.service.mail.text.une.content", args);
         } else {
-            String inDirectory = basicServiceForm.isInDirectory() ? getResources(request).getMessage("select.yes") : getResources(request).getMessage("select.no");
-            String[] args = {basicServiceForm.getUser(), basicServiceForm.getDomain(), basicServiceForm.getProfundidad(), basicServiceForm.getAmplitud(), inDirectory, basicServiceForm.getReport()};
+            final String inDirectory = basicServiceForm.isInDirectory() ? getResources(request).getMessage("select.yes") : getResources(request).getMessage("select.no");
+            final String[] args = {basicServiceForm.getUser(), basicServiceForm.getDomain(), basicServiceForm.getProfundidad(), basicServiceForm.getAmplitud(), inDirectory, reportToString(basicServiceForm.getReport())};
             text = basicServiceForm.getReport().equals(Constants.REPORT_OBSERVATORY) ? getResources(request).getMessage("basic.service.mail.text.observatory", args) : getResources(request).getMessage("basic.service.mail.text.une", args);
         }
 
         return text;
+    }
+
+    private String reportToString(final String reportType) {
+        if (Constants.REPORT_OBSERVATORY.equals(reportType) || Constants.REPORT_OBSERVATORY_FILE.equals(reportType)) {
+            return "Observatorio UNE 2004";
+        } else if (Constants.REPORT_OBSERVATORY_2.equals(reportType)) {
+            return "Observatorio UNE 2012";
+        } else if ("observatorio-1-nobroken".equals(reportType)) {
+            return "Observatorio UNE 2004 (sin comprobar enlaces rotos)";
+        } else if (Constants.REPORT_OBSERVATORY_2_NOBROKEN.equals(reportType)) {
+            return "Observatorio UNE 2012 (sin comprobar enlaces rotos)";
+        } else {
+            return reportType;
+        }
     }
 
 }
